@@ -1,15 +1,15 @@
 source 'https://rubygems.org'
-ruby '2.0.0'
+ruby '2.2.2'
 
-gem 'bootstrap-sass', '~> 2.0.4.0'
-gem 'coffee-rails', '~> 4.0.0'
+gem 'bootstrap-sass'
+gem 'coffee-rails'
 gem 'draper'
 gem 'foreman'
-gem 'jbuilder', '~> 1.0.1'
+gem 'jbuilder'
 gem 'jquery-rails'
 gem 'pg'
-gem 'rails', '4.0.0.rc1'
-gem 'sass-rails', '~> 4.0.0.rc1'
+gem 'rails'
+gem 'sass-rails'
 gem 'slim-rails'
 gem 'sidekiq'
 gem 'sinatra'
@@ -26,6 +26,7 @@ group :test, :development do
   gem 'factory_girl_rails'
   gem 'forgery'
   gem 'zonebie'
+  gem 'pry'
 end
 
 group :test do
