@@ -35,23 +35,42 @@ module LaundryView
       doc = Nokogiri::HTML body
       stats = doc.css('span.stat')
       machines = stats.map {|s| LaundryView::Machine.new(s.text.strip)}
+      types = doc.css('img')
+      machines.each_with_index do |m, i|
+        test = types[i*2].attributes["src"].value
+        if test =~ /washer/
+          m.kind = "washer"
+        elsif test =~ /dryer/
+          m.kind = "dryer"
+        end
+      end
+      machines
     end
   end
 
   class Machine
+    attr_reader :remaining, :broken
+    attr_accessor :kind
+
     def initialize status
-      @status = status
+      if status == "out of service"
+        @broken = true
+        @remaining = 0
+      else
+        @broken = false
+        @remaining = calc_remaining(status)
+      end
     end
 
-    def status
-      @status
-    end
+    private
 
-    def remaining
-      return 0 if @status == "cycle has ended - door still closed"
-      return -1 if @status == "available"
-      if @status =~ /est. time remaining [0-9]+ min/ or @status =~ /running for [0-9]+ mins/
-        return @status.match(/[0-9]+/)[0].to_i
+    def calc_remaining status
+      if status == "cycle has ended - door still closed"
+        return 0
+      end
+      return nil if status == "available"
+      if status =~ /est. time remaining [0-9]+ min/ or status =~ /running for [0-9]+ mins/
+        return status.match(/[0-9]+/)[0].to_i
       end
       9999
     end
